Changes regarding the cadastre redefinition.
In geotools there is introduced an abstract tool which draws a rectangle in the screen.
In gis project a boundle to handle cadastre redefinition has been defined along with a layer for modified cadastre objects.
Added in cadastre ejb a method to retrieve a node and the partecipating cadastre objects.