Make historic parcel searchable [#74 state:fixed responsible:none]
It was: when searched for cadastre objects by a ba_unit (title), are
returned only the cadastre objects that are current. When the cadastre
object goes historic (from a cadastre change) although the title is still
current, the cadastre object is not found any more.
This is changed to make cadastre objects searchable that are current or
they have a least one title still current.
The same is done also for the search by owner in the title.
This rises another issue: At the moment, in the map are displayed only
current parcels. This kind of search, is actually searching for parcels
that are not displayed and can bring confusion.