Set default values for baUnit
The default value for baUnit has to be basicPropertyUnit.
From the code is removed and it is set in the database level as many other
default values.
Also the status_code is set to pending in the database level.
To make this run, you need also to rebuild the database.