Consider map resolution in getSpatial [#68]
In the searchEJB, the method that retrieves features for navigation in the
map considers an extra parameter called pixel_res which is the pixel
resolution at that moment in the map.
As needed, this paramter can be used in queries to reduce number of
features for a certain area and resolution.