Fix approve cadastre object service issue [#219 state:fixed responsible:none]
The query that is used to retrieve the pending cadastre objects was
retrieving also current objects. So the current objects were twice changed
and because the changing of the status of pending to current can bring
changes to existing other current objects in case new points have been
introduced, this makes the rowversion of the current to go more up and it
was a difference between the rowversion of the record in database and the
one selected before.