Fix search for LOT 47 Plan 10477
Fixed in v2.1.4. This issue is caused because the search
splits the search string into mulitple word parts using
the space character has the word delimiter.

To overcome this problem, I have updated the search to
recognize and retain the POSIX escape character for space (i.e. \s).
This will allow the user to explicitly indicate where a
space should occur in the search string and match based
on that. E.g. to search for LOT 47 in plan 10477 the
search string to use is LOT\s47 10477

[#89 state:fixed responsible:none milestone:v2.1.4]