Fix for Change Estate Type service
Fixed in v2.1.4. The correct way to change the estate type
for a property is to use the Change Estate Type service,
unfortunately this did have a bug which has now been fixed.
I have also updated the Change Estate Type service to have
a 0 fee as it will likely be part of another service
such as a Court Order.
[#92 state:fixed responsible:none milestone:v2.1.4]