Folio Certificate for Unit Titles [#67 responsible:none state:fixed]
Created Folio Certificate for Unit Titles based on the existing
Computer Folio Certificate. The Unit Titles certificate presents
the required information, but it is likely to require further
customization as part of the process of preparing for the first
official Unit Plan.  In the meantime, the Training Report
watermark will always display whether the system is production
or training.